+++ kevlar/__init__.py
@@ -29,6 +29,8 @@ from kevlar import seqio
 from kevlar import overlap
 from kevlar import sketch
 from kevlar import reference
+from kevlar import varmap
+from kevlar import vcf
 from kevlar.mutablestring import MutableString
 from kevlar.readgraph import ReadGraph
 from kevlar.seqio import parse_augmented_fastx, print_augmented_fastx