+++ kevlar/tests/test_call.py
@@ -7,14 +7,13 @@
 # licensed under the MIT license: see LICENSE.
 # -----------------------------------------------------------------------------
 
-from io import StringIO
-import pytest
-import re
+
 import sys
 import khmer
 import kevlar
-from kevlar.call import call, alignment_interpretable, VariantMapping
+from kevlar.call import call
 from kevlar.tests import data_file
+import pytest
 import screed
 
 
@@ -28,17 +27,6 @@ def test_align():
              'TTCAAGCGCAGGTTCGAGCCAGTCAGGACTGCTCCCCCACTTCCTCAAGTCTCATCGTGTTTTT'
              'TTTAGAGCTAGTTTCTTAGTCTCATTAGGCTTCAGTCACCATCATTTCTTATAGGAATACCA')
     assert kevlar.align(target, query) == ('10D91M69D79M20I', 155)
-
-
-def test_cigar_scrutability():
-    assert alignment_interpretable('6M73I13M4I5M4D63M11I') is False
-    assert alignment_interpretable('29I17M7I3M24I36M10D7M11I14M23I4M') is False
-    assert alignment_interpretable('57I16M7I3M8D33M1I28M27I3M') is False
-
-    assert alignment_interpretable('25D100M25D') is True
-    assert alignment_interpretable('50I50M50I50M24D1M') is True
-    assert alignment_interpretable('47I127M30D1M') is True
-    assert alignment_interpretable('30D100M16D67M30D') is True
 
 
 @pytest.mark.parametrize('ccid,varcall', [
@@ -66,8 +54,7 @@ def test_call_pico_indel(ccid, varcall):
     ('223', '50D268M50D1M', '5:42345359:C->G'),
 ])
 def test_call_ssc_isolated_snv(ccid, cigar, varcall):
-    """
-    Ensure isolated SNVs are called correctly.
+    """Ensure isolated SNVs are called correctly.
 
     SNVs that are well separated from other variants have a distinct alignment
     signature as reflected in the CIGAR string reported by ksw2. They are
@@ -87,44 +74,6 @@ def test_call_ssc_isolated_snv(ccid, cigar, varcall):
     assert str(calls[0]) == varcall
 
 
-def test_call_ssc_1bpdel():
-    """Test 1bp deletion"""
-    qfile = data_file('ssc218.contig.augfasta')
-    tfile = data_file('ssc218.gdna.fa')
-
-    qinstream = kevlar.parse_augmented_fastx(kevlar.open(qfile, 'r'))
-    query = [record for record in qinstream][0]
-    tinstream = kevlar.reference.load_refr_cutouts(kevlar.open(tfile, 'r'))
-    target = [record for record in tinstream][0]
-    aln = VariantMapping(query, target, 1e6, '50D132M1D125M50D')
-    variants = aln.call_variants(31)
-
-    assert isinstance(variants, list)
-    assert len(variants) == 1
-    assert str(variants[0]) == '6:23230160:1D'
-
-
-def test_call_ssc_two_proximal_snvs():
-    """
-    Test two proximal SNVs
-
-    Currently this serves as a negative control for calling isolated SNVs, but
-    distinguishing which (if any) of a set of proximal SNVs is novel will be
-    supported soon, and this test will need to be updated.
-    """
-    qfile = data_file('ssc107.contig.augfasta.gz')
-    tfile = data_file('ssc107.gdna.fa.gz')
-
-    qinstream = kevlar.parse_augmented_fastx(kevlar.open(qfile, 'r'))
-    query = [record for record in qinstream][0]
-    tinstream = kevlar.reference.load_refr_cutouts(kevlar.open(tfile, 'r'))
-    target = [record for record in tinstream][0]
-
-    aln = VariantMapping(query, target, 1e6, '25D263M25D')
-    variants = aln.call_variants(31)
-    assert len(variants) == 2
-
-
 @pytest.mark.parametrize('targetfile,queryfile,cigar', [
     ('pico-7-refr.fa', 'pico-7-asmbl.fa', '10D83M190D75M20I1M'),
     ('pico-2-refr.fa', 'pico-2-asmbl.fa', '10D89M153I75M20I'),
@@ -138,39 +87,6 @@ def test_call_formerly_inscrutable(targetfile, queryfile, cigar, capsys):
     out, err = capsys.readouterr()
     print(out)
     assert 'GC=' not in out
-
-
-def test_snv_obj():
-    snv = kevlar.call.VariantSNV('scaffold42', 10773, 'A', 'G')
-    assert str(snv) == 'scaffold42:10773:A->G'
-    vcfvalues = ['scaffold42', '10774', '.', 'A', 'G', '.', 'PASS', '.']
-    assert snv.vcf == '\t'.join(vcfvalues)
-    assert snv.cigar is None
-
-    snv2 = kevlar.call.VariantSNV('chr5', 500, 'T', 'G', CG='10D200M10D')
-    assert snv2.cigar == '10D200M10D'
-    assert snv2.window is None
-
-
-def test_indel_obj():
-    """
-    Test indel objects
-
-    The coordinate used to construct the object is 0-based, but includes the
-    nucleotide shared by the reference and alternate alleles. The str() output
-    coordinate is increased by 1 to account for this nucleotide, while the VCF
-    output is increased by 1 to transform to a 1-based system where the shared
-    nucleotide is the point of reference.
-    """
-    indel1 = kevlar.call.VariantIndel('chr3', 8998622, 'GATTACA', 'G')
-    assert str(indel1) == 'chr3:8998623:6D'
-    vcfvalues = ['chr3', '8998623', '.', 'GATTACA', 'G', '.', 'PASS', '.']
-    assert indel1.vcf == '\t'.join(vcfvalues)
-
-    indel2 = kevlar.call.VariantIndel('chr6', 75522411, 'G', 'GATTACA')
-    assert str(indel2) == 'chr6:75522412:I->ATTACA'
-    vcfvalues = ['chr6', '75522412', '.', 'G', 'GATTACA', '.', 'PASS', '.']
-    assert indel2.vcf == '\t'.join(vcfvalues)
 
 
 def test_variant_kmers():
@@ -190,71 +106,14 @@ def test_variant_kmers():
     assert calls[0].window == window
 
 
-@pytest.mark.parametrize('prefix,cigar,refrwindow,altwindow', [
-    ('phony-snv-01', '25D98M25D',
-        'GGGGGTGTCTGCGACCACAGCTGAACATGACGAAACGGGTG',
-        'GGGGGTGTCTGCGACCACAGGTGAACATGACGAAACGGGTG'),
-    ('phony-snv-02', '24D99M25D',
-        'ATTCGTATTACCCCTGGGATTTGGGAGCTGGTCTATATAGG',
-        'ATTCGTATTACCCCTGGGATATGGGAGCTGGTCTATATAGG'),
-    ('phony-deletion-01', '25D28M8D49M25D',
-        'GGCTCAAGACTAAAAAGACTGAGACTCGTTTTTGGTGACAAGCAGGGC',
-        'GGCTCAAGACTAAAAAGACTTTTTTGGTGACAAGCAGGGC'),
-    ('phony-deletion-02', '40D29M3D36M40D',
-        'CATCATCTCGTAGGTTTGTCTAGTGCAAACAGAGTCCCCCTGC',
-        'CATCATCTCGTAGGTTTGTCTGCAAACAGAGTCCCCCTGC'),
-    ('phony-insertion-01', '10D34M7I49M10D1M',
-        'CATCTGTTTTTCTCGAACTCGTATATTATCTATAAATTCC',
-        'CATCTGTTTTTCTCGAACTCGATTACAGTATATTATCTATAAATTCC'),
-    ('phony-insertion-02', '10D33M27I95M10D',
-        'GCCAGGAAGTTTACGATAAGGTGTTGCCATTCGAAATGAC',
-        'GCCAGGAAGTTTACGATAAGTATATATATATATATATATATATATATGTGTTGCCATTCGAAATGAC'),
-])
-def test_variant_window(prefix, cigar, refrwindow, altwindow):
-    qfile = data_file(prefix + '.contig.fa')
-    tfile = data_file(prefix + '.gdna.fa')
-
-    qinstream = kevlar.parse_augmented_fastx(kevlar.open(qfile, 'r'))
-    query = [record for record in qinstream][0]
-    tinstream = kevlar.reference.load_refr_cutouts(kevlar.open(tfile, 'r'))
-    target = [record for record in tinstream][0]
-
-    aln = VariantMapping(query, target, 1e6, cigar)
-    variants = aln.call_variants(21)
-    assert len(variants) == 1
-    assert variants[0].window == altwindow
-    assert variants[0].refrwindow == refrwindow
-
-
-def test_nocall():
-    # Intentionally mismatched
-    qfile = data_file('phony-deletion-01.contig.fa')
-    tfile = data_file('phony-insertion-01.gdna.fa')
-
-    qinstream = kevlar.parse_augmented_fastx(kevlar.open(qfile, 'r'))
-    query = [record for record in qinstream][0]
-    tinstream = kevlar.reference.load_refr_cutouts(kevlar.open(tfile, 'r'))
-    target = [record for record in tinstream][0]
-
-    aln = VariantMapping(query, target, 1e6, '25D5M22I5M46D8M13D2M35I')
-    variants = aln.call_variants(21)
-    assert len(variants) == 1
-    assert variants[0].vcf == (
-        'yourchr\t801\t.\t.\t.\t.\t.\t'
-        'CG=25D5M22I5M46D8M13D2M35I;NC=inscrutablecigar;QN=contig4:cc=1;'
-        'QS=AACTGGTGGGCTCAAGACTAAAAAGACTTTTTTGGTGACAAGCAGGGCGGCCTGCCCTTCCTGTAG'
-        'TGCAAGAAAAT'
-    )
-
-
 @pytest.mark.parametrize('part,coord,window', [
-    (12, 7027071, 'CAGGGAGAGGCAGCCTGCCCTCAACCTGGGAGAGCACTGTCTAATCAGCTCCCATCTCA'
+    (12, 7027087, 'CAGGGAGAGGCAGCCTGCCCTCAACCTGGGAGAGCACTGTCTAATCAGCTCCCATCTCA'
                   'GG'),
     (16, 25755121, 'TTTTGGTGTTTAGACATGAAGTCCTTGCCCATCGAGTTATGCCTATGTCCTGAATGCT'
                    'ATTGCCTAGG'),
     (23, 59459928, 'CAGGCGTGAGCCACCGCGCCTGGCCAGGAGCATTGTTTGAACCCAGAAGGCGGAGGTT'
                    'GCA'),
-    (192, 28556906, 'AAAATACAAAAATTAGCCAGGCATGGTGGTGCATGCCTGTAATACCAGCCTTTTAGA'
+    (192, 28556953, 'AAAATACAAAAATTAGCCAGGCATGGTGGTGCATGCCTGTAATACCAGCCTTTTAGA'
                     'GGC')
 ])
 def test_funky_cigar(part, coord, window):
@@ -268,6 +127,7 @@ def test_funky_cigar(part, coord, window):
 
     calls = list(call(targets, contigs))
     assert len(calls) == 1
+    print('DEBUG', calls[0].vcf, file=sys.stderr)
     assert calls[0].seqid == '17'
     assert calls[0].position == coord - 1
     assert calls[0].info['VW'] == window
@@ -327,21 +187,6 @@ def test_multibest_revcom():
                             'GCCTGAGCCCCA')
 
 
-def test_variant_mapping():
-    contig = screed.Record(
-        name='contig1',
-        sequence='CCTGAGCCCTCTCAAGTCGGGTCCTGGCCCGGTCTGCCCATGAGGCTGGGCCTGAGCCCC'
-    )
-    cutout = kevlar.reference.ReferenceCutout(
-        defline='chr1_10000-10060',
-        sequence='CCTGAGCCCTCTCAAGTCGGGTCCTGGCCCAGTCTGCCCATGAGGCTGGGCCTGAGCCCC'
-    )
-    mapping = VariantMapping(contig, cutout, score=1e6, cigar='60M')
-
-    assert mapping.seqid == 'chr1'
-    assert mapping.interval == ('chr1', 10000, 10060)
-
-
 def test_align_mates():
     mate_seqs = kevlar.open(data_file('minitrio/novel-mates.fastq.gz'), 'r')
     record = screed.Record(
@@ -373,89 +218,3 @@ def test_mate_distance():
     positions = [('seq2', 4000), ('seq2', 3000), ('seq2', 5100), ('seq3', 1)]
     gdna_pos = ('seq2', 5000, 5500)
     assert kevlar.call.mate_distance(positions, gdna_pos) == 1000.0
-
-
-@pytest.mark.parametrize('query,target,dist,n,msgcount', [
-    ('phony-snv-01b.contig.fa', 'phony-snv-01.gdna.fa', 5, 1, 1),
-    ('phony-snv-02b.contig.fa', 'phony-snv-02.gdna.fa', 5, 1, 1),
-    ('phony-snv-01b.contig.fa', 'phony-snv-01.gdna.fa', 2, 2, 0),
-    ('phony-snv-02b.contig.fa', 'phony-snv-02.gdna.fa', None, 2, 0),
-])
-def test_call_near_end(query, target, dist, n, msgcount):
-    log = StringIO()
-    contig = next(
-        kevlar.parse_augmented_fastx(
-            kevlar.open(data_file(query), 'r')
-        )
-    )
-    cutout = next(
-        kevlar.reference.load_refr_cutouts(
-            kevlar.open(data_file(target), 'r')
-        )
-    )
-    aln = kevlar.call.align_both_strands(cutout, contig)
-    calls = list(aln.call_variants(31, mindist=dist, logstream=log))
-    assert len(calls) == n
-
-    err = log.getvalue()
-    count = err.count('discarding SNV due to proximity to end of the contig')
-    assert count == msgcount
-
-
-@pytest.mark.parametrize('query,target,vw,rw', [
-    (
-        'trunc-snv.contig.fa', 'trunc-snv.gdna.fa',
-        'TAGCATACAGGTAGTCAGGGGGTGTCTGCGACCACAGCTGAA',
-        'TAGCATACAGGAAGTCAGGGGGTGTCTGCGACCACAGCTGAA'
-    ),
-    (
-        'trunc-indel.contig.augfasta', 'trunc-indel.gdna.augfasta',
-        'GTATATACACATATATGTGTATATATGTGTATATATGT',
-        'GTATATACATATATGTGTATATATGTGTATATACATATATGTGTATATATGTGTATATATGT'
-    ),
-    (
-        'trunc-snv-funky.contig.fa', 'trunc-snv-funky.gdna.fa',
-        'TGTGTCTGAGAGGGTGTTGCCAAAGGAGATTAACATTTG',
-        'TGTGTCTGTGAGGGTGTTGCCAAAGGAGATTAACATTTG'
-    ),
-    (
-        'trunc-indel-funky.contig.fa', 'trunc-snv-funky.gdna.fa',
-        'TGTGTCTGTGAGTATATAGGTGTTGCCAAAGGAGATTAACATTTGAGT',
-        'TGTGTCTGTGAGGGTGTTGCCAAAGGAGATTAACATTTGAGT'
-    ),
-])
-def test_call_truncated_windows(query, target, vw, rw):
-    contig = next(
-        kevlar.parse_augmented_fastx(
-            kevlar.open(data_file(query), 'r')
-        )
-    )
-    cutout = next(
-        kevlar.reference.load_refr_cutouts(
-            kevlar.open(data_file(target), 'r')
-        )
-    )
-    aln = kevlar.call.align_both_strands(cutout, contig)
-    calls = list(aln.call_variants(31))
-    assert len(calls) == 1
-    print('VW:', calls[0].window, file=sys.stderr)
-    print('RW:', calls[0].refrwindow, file=sys.stderr)
-    assert calls[0].window == vw
-    assert calls[0].refrwindow == rw
-
-
-def test_call_num_interesting_kmers():
-    contig = next(
-        kevlar.parse_augmented_fastx(
-            kevlar.open(data_file('iktest.contig.fa'), 'r')
-        )
-    )
-    cutout = next(
-        kevlar.reference.load_refr_cutouts(
-            kevlar.open(data_file('iktest.gdna.fa'), 'r')
-        )
-    )
-    aln = kevlar.call.align_both_strands(cutout, contig)
-    calls = list(aln.call_variants(29))
-    assert len(calls) == 1
-    assert calls[0].info['IK'] == '1'